+++ include/ccf/endpoint_registry.h
@@ -39,6 +39,25 @@ namespace ccf::endpoints
     PathTemplateSpec spec;
   };
 
+  struct RequestCompletedEvent
+  {
+    std::string method = "";
+    // This contains the path template against which the request matched. For
+    // instance `/user/{user_id}` rather than `/user/Bob`. This should be safe
+    // to log, though doing so still reveals (to anyone with stdout access)
+    // exactly which endpoints were executed and when.
+    std::string dispatch_path = "";
+    int status = 0;
+    std::chrono::milliseconds exec_time{0};
+    size_t attempts = 0;
+  };
+
+  struct DispatchFailedEvent
+  {
+    std::string method = "";
+    int status = 0;
+  };
+
   void default_locally_committed_func(
     CommandEndpointContext& ctx, const TxID& tx_id);
 
@@ -113,14 +132,6 @@ namespace ccf::endpoints
       std::string document_version = "0.0.1";
     } openapi_info;
 
-    struct Metrics
-    {
-      size_t calls = 0;
-      size_t errors = 0;
-      size_t failures = 0;
-      size_t retries = 0;
-    };
-
     template <typename T>
     bool get_path_param(
       const ccf::PathParams& params,
@@ -150,12 +161,6 @@ namespace ccf::endpoints
       std::string,
       std::map<RESTVerb, std::shared_ptr<PathTemplatedEndpoint>>>
       templated_endpoints;
-
-    ccf::pal::Mutex metrics_lock;
-    std::map<std::string, std::map<std::string, Metrics>> metrics;
-
-    EndpointRegistry::Metrics& get_metrics_for_request(
-      const std::string& method, const std::string& verb);
 
     kv::Consensus* consensus = nullptr;
     kv::TxHistory* history = nullptr;
@@ -278,13 +283,14 @@ namespace ccf::endpoints
 
     void set_history(kv::TxHistory* h);
 
-    virtual void increment_metrics_calls(const EndpointDefinitionPtr& endpoint);
-    virtual void increment_metrics_errors(
-      const EndpointDefinitionPtr& endpoint);
-    virtual void increment_metrics_failures(
-      const EndpointDefinitionPtr& endpoint);
-    virtual void increment_metrics_retries(
-      const EndpointDefinitionPtr& endpoint);
+    // Override these methods to log or report request metrics.
+    virtual void handle_event_request_completed(
+      const ccf::endpoints::RequestCompletedEvent& event)
+    {}
+
+    virtual void handle_event_dispatch_failed(
+      const ccf::endpoints::DispatchFailedEvent& event)
+    {}
 
     virtual bool apply_uncommitted_tx_backpressure() const
     {