+++ include/ccf/research/custom_protocol_subsystem_interface.h
@@ -12,19 +12,19 @@
 #include <functional>
 #include <memory>
 
-namespace tls
-{
-  class Context;
-  using ConnID = int64_t;
-}
-
 namespace ccf
 {
+  namespace tls
+  {
+    class Context;
+    using ConnID = int64_t;
+  }
+
   class CustomProtocolSubsystemInterface : public AbstractNodeSubSystem
   {
   public:
     using CreateSessionFn = std::function<std::shared_ptr<Session>(
-      tls::ConnID, const std::unique_ptr<tls::Context>&&)>;
+      ccf::tls::ConnID, const std::unique_ptr<tls::Context>&&)>;
 
     virtual ~CustomProtocolSubsystemInterface() = default;
 
@@ -40,7 +40,7 @@ namespace ccf
 
     virtual std::shared_ptr<Session> create_session(
       const std::string& protocol_name,
-      tls::ConnID conn_id,
+      ccf::tls::ConnID conn_id,
       const std::unique_ptr<tls::Context>&& ctx) = 0;
 
     struct Essentials