+++ src/ds/test/hex.cpp
@@ -17,17 +17,17 @@ TEST_CASE("Hex string to and from conversion")
     INFO("Simple test");
 
     std::string valid_hex_str("0123456789abcdef");
-    auto data = ds::from_hex(valid_hex_str);
+    auto data = ccf::ds::from_hex(valid_hex_str);
     REQUIRE(data.size() == valid_hex_str.size() / 2);
-    REQUIRE(ds::to_hex(data) == valid_hex_str);
+    REQUIRE(ccf::ds::to_hex(data) == valid_hex_str);
   }
 
   {
     INFO("Invalid length");
 
-    REQUIRE_THROWS(ds::from_hex("a"));
-    REQUIRE_THROWS(ds::from_hex("abc"));
-    REQUIRE_THROWS(ds::from_hex("abcde"));
+    REQUIRE_THROWS(ccf::ds::from_hex("a"));
+    REQUIRE_THROWS(ccf::ds::from_hex("abc"));
+    REQUIRE_THROWS(ccf::ds::from_hex("abcde"));
   }
 
   {
@@ -37,8 +37,8 @@ TEST_CASE("Hex string to and from conversion")
     std::string lowercase_hex_str = uppercase_hex_str;
     nonstd::to_lower(lowercase_hex_str);
 
-    auto data = ds::from_hex(uppercase_hex_str);
+    auto data = ccf::ds::from_hex(uppercase_hex_str);
     REQUIRE(data.size() == uppercase_hex_str.size() / 2);
-    REQUIRE(ds::to_hex(data) == lowercase_hex_str);
+    REQUIRE(ccf::ds::to_hex(data) == lowercase_hex_str);
   }
 }