+++ src/enclave/session.h
@@ -26,7 +26,7 @@ namespace ccf
     ThreadedSession(int64_t thread_affinity)
     {
       execution_thread =
-        threading::ThreadMessaging::instance().get_execution_thread(
+        ::threading::ThreadMessaging::instance().get_execution_thread(
           thread_affinity);
     }
 
@@ -34,19 +34,19 @@ namespace ccf
     // that eventually invokes the virtual handle_incoming_data_thread()
     void handle_incoming_data(std::span<const uint8_t> data) override
     {
-      auto [_, body] = ringbuffer::read_message<tls::tls_inbound>(data);
+      auto [_, body] = ringbuffer::read_message<::tls::tls_inbound>(data);
 
-      auto msg = std::make_unique<threading::Tmsg<SendRecvMsg>>(
+      auto msg = std::make_unique<::threading::Tmsg<SendRecvMsg>>(
         &handle_incoming_data_cb);
       msg->data.self = this->shared_from_this();
       msg->data.data.assign(body.data, body.data + body.size);
 
-      threading::ThreadMessaging::instance().add_task(
+      ::threading::ThreadMessaging::instance().add_task(
         execution_thread, std::move(msg));
     }
 
     static void handle_incoming_data_cb(
-      std::unique_ptr<threading::Tmsg<SendRecvMsg>> msg)
+      std::unique_ptr<::threading::Tmsg<SendRecvMsg>> msg)
     {
       msg->data.self->handle_incoming_data_thread(std::move(msg->data.data));
     }